requirements.txt
streamlit>=1.30.0
fastapi>=0.100.0
uvicorn>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openai>=1.0.0
faiss-cpu>=1.7.4
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
scikit-learn>=1.3.0
folium>=0.14.0
streamlit-folium>=0.15.0
networkx>=3.0
gremlinpython>=3.7.0
